+++ nilearn/input_data/nifti_masker.py
@@ -34,7 +34,8 @@ def filter_and_mask(imgs, mask_img_, parameters,
                     confounds=None,
                     copy=True,
                     dtype=None):
-    imgs = _utils.check_niimg(imgs, atleast_4d=True, ensure_ndim=4)
+    imgs = _utils.check_niimg(imgs, atleast_4d=True, ensure_ndim=4,
+                              dtype=dtype)
 
     # Check whether resampling is truly necessary. If so, crop mask
     # as small as possible in order to speed up the process
